requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.21.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.25.0
datasets>=2.13.0
huggingface-hub>=0.16.0
tqdm>=4.66.0